Move Examples to `example` Package
Help to differentiate example code from Caliban library code by disentangling the packages.
