+++ examples/src/main/scala/example/ExampleApi.scala
@@ -1,0 +1,56 @@
+package example
+
+import example.ExampleData._
+import example.ExampleService.ExampleService
+
+import caliban.GraphQL
+import caliban.GraphQL.graphQL
+import caliban.RootResolver
+import caliban.schema.Annotations.{ GQLDeprecated, GQLDescription }
+import caliban.schema.GenericSchema
+import caliban.wrappers.ApolloTracing.apolloTracing
+import caliban.wrappers.Wrappers._
+
+import zio.URIO
+import zio.clock.Clock
+import zio.console.Console
+import zio.duration._
+import zio.stream.ZStream
+
+import scala.language.postfixOps
+
+object ExampleApi extends GenericSchema[ExampleService] {
+
+  case class Queries(
+    @GQLDescription("Return all characters from a given origin")
+    characters: CharactersArgs => URIO[ExampleService, List[Character]],
+    @GQLDeprecated("Use `characters`")
+    character: CharacterArgs => URIO[ExampleService, Option[Character]]
+  )
+  case class Mutations(deleteCharacter: CharacterArgs => URIO[ExampleService, Boolean])
+  case class Subscriptions(characterDeleted: ZStream[ExampleService, Nothing, String])
+
+  implicit val roleSchema           = gen[Role]
+  implicit val characterSchema      = gen[Character]
+  implicit val characterArgsSchema  = gen[CharacterArgs]
+  implicit val charactersArgsSchema = gen[CharactersArgs]
+
+  val api: GraphQL[Console with Clock with ExampleService] =
+    graphQL(
+      RootResolver(
+        Queries(
+          args => ExampleService.getCharacters(args.origin),
+          args => ExampleService.findCharacter(args.name)
+        ),
+        Mutations(args => ExampleService.deleteCharacter(args.name)),
+        Subscriptions(ExampleService.deletedEvents)
+      )
+    ) @@
+      maxFields(200) @@               // query analyzer that limit query fields
+      maxDepth(30) @@                 // query analyzer that limit query depth
+      timeout(3 seconds) @@           // wrapper that fails slow queries
+      printSlowQueries(500 millis) @@ // wrapper that logs slow queries
+      printErrors @@                  // wrapper that logs errors
+      apolloTracing                   // wrapper for https://github.com/apollographql/apollo-tracing
+
+}
