+++ project/ConsoleHelper.scala
@@ -20,8 +20,8 @@ object ConsoleHelper {
             |${item("test")} - Run the unit test suite
             |${item("fmt")} - Run scalafmt on the entire project
             |${item("scripted")} - Run the scripted test suite
-            |${item("examples/runMain caliban.http4s.ExampleApp")} - Start the example server (http4s based)
-            |${item("examples/runMain caliban.akkahttp.ExampleApp")} - Start the example server (akka-http based)
+            |${item("examples/runMain example.http4s.ExampleApp")} - Start the example server (http4s based)
+            |${item("examples/runMain example.akkahttp.ExampleApp")} - Start the example server (akka-http based)
             |${item("benchmarks/jmh:run")} - Run the benchmarks
             |${item("+publishLocal")} - Publish caliban locally""".stripMargin
